Renaming modules, simplifying code
