+++ src/de/mod.rs
@@ -1,4 +1,4 @@
-pub mod deserialize;
 mod deserializer;
 mod packet_reader;
+pub mod packets;
 pub(crate) use packet_reader::PacketReader;
